refactor(update): BaseMap 패키지 변경
